serial_monitor_app/requirements.txt
pyserial
PySide6
matplotlib
numpy